python-service/requirements.txt
# 基你太美 - Python AI服务依赖包

# Web框架
fastapi==0.108.0
uvicorn[standard]==0.25.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 数据处理
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# AI和机器学习
openai==1.6.1
dashscope==1.17.0
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# 图像处理和OCR
Pillow==10.1.0
opencv-python==4.8.1.78
tencentcloud-sdk-python==3.0.1004

# 数据验证
pydantic==2.5.2
pydantic-settings==2.1.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 工具类
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 异步支持
asyncio==3.4.3

# 开发工具
black==23.12.1
flake8==7.0.0
pytest==7.4.3
pytest-asyncio==0.21.1

# 其他工具
requests==2.31.0
beautifulsoup4==4.12.2
jieba==0.42.1